leadgen_api_v1/requirements.txt
Django==2.0
djangorestframework==3.7.7
requests==2.18.4
timezonefinder==2.1.2
